update UI
Now that restrictions have loosened around the FAT fs (namely that
OS/folders are no longer hard-coded), an update to the UI is needed.
These directory names are no longer guaranteed, and we need to generate
the page dynamically. Additionally, some scripting was added to format
the webpage into raw bytes to be included into the source as PROGMEM.